List of read packets owned by TrasportStream
The list of packets are now owned by TransportStream and shared with
FileSource. FileSource adds packets to it and signals when it have
enough packets. TSReport calls TransportStream::getPackets which waits
for signal from FileSource. When signal is received the list is copied
and returned. The shared list is then cleared.

Does not yet work. The entire file is not read, hangs some time.